+++ resources/views/components/money.blade.php
@@ -1,11 +1,3 @@
 @props(['amount', 'currency' => null])
 
-@php
-    $currency ??= config('app.currency', 'GHS');
-    $symbol = match ($currency) {
-        'GHS' => 'GH₵',
-        default => $currency.' ',
-    };
-@endphp
-
-{{ $symbol.number_format(($amount ?? 0) / 100, 2) }}
+{{ \App\Support\CurrencySymbol::for($currency).number_format(($amount ?? 0) / 100, 2) }}